arch2_microservices/book_service/requirements.txt
grpcio
grpcio-tools

sqlalchemy
psycopg2-binary
alembic